video_converter/requirements.txt
streamlit
google-cloud-speech
google-cloud-texttospeech
moviepy
requests
